GMSTR-4: added orderService logic and tests for it

+++ GameShop.BLL/DTO/OrderDetailsDTOs/OrderDetailsCreateDTO.cs
@@ -10,6 +10,6 @@ namespace GameShop.BLL.DTO.OrderDetailsDTOs
     {
         public string GameKey { get; set; }
 
-        public int Quantity { get; set; }
+        public short Quantity { get; set; }
     }
 }
